refactor(AudioAsset): use "V2" instead of "New" for the new version of components
