+++ src/script/components/MessagesList/Message/ContentMessage/asset/AudioAsset/AudioAssetV2.styles.ts
